source/requirements.txt
numpy==1.26.4
matplotlib==3.9.2
opencv==4.10.0
pathlib==1.0.1
pydot==3.0.2
pyyaml==6.0.2
tensorflow==2.17.0
seaborn==0.13.2